Finished functions.
The last 2 bent my mind for a while. Hope I got them right. You are
given a form which has a blank at the beginning and an anonymous
'square' function at the end. So, I figure I need too create a higher
order function whose parameter is a square function. That anonymous
function should call it's parameter with the value 5.